Some basic maintenance to conclude the refresh of `creative` : 

- Merge two conditions into one.
- Simplify a Lua pattern.
- New (small) feature : go to the last page on clicking on [<<] at page 1.