@paramat 2015-12-26
```
	if not (def.groups.not_in_creative_inventory == 1) and
				def.description and def.description ~= ""
```
Is simple and fast while:
```
if (filter and def.name:find(filter, 1, true)) or not filter
```
.. looks to be a slower operation, so having 2 'if' statements might actually be the optimum.

Making the inventory pages wrap is a good feature i have been wanting.