@kilbith 2015-12-26
Please stop these pointless comments on the `if` statements.

I have done some benchmarks and got an equal average on this piece of code : 
```C
Before : 0.65 ms
Before : 1.13 ms

After : 0.34 ms
After : 0.75 ms
```
It's processed the same way than before with no impact on speed.