@rubenwardy 2015-12-26
> .. looks to be a slower operation, so having 2 'if' statements might actually be the optimum.

If you have an if statement like so:

```lua
if false and foo and bar then
```

Then foo and bar will never be executed. It's like this in pretty much every language.