@paramat 2015-12-27
Thanks rubenwardy i see.
The question was valid not pointless.